+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/messages/MecanumCommandMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team26396.teamcode.messages;
+package org.firstinspires.ftc.team26396.roadrunner.messages;
 
 public final class MecanumCommandMessage {
     public long timestamp;
