+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/tuning/SplineTest.java
@@ -1,12 +1,12 @@
-package org.firstinspires.ftc.team26396.teamcode.tuning;
+package org.firstinspires.ftc.team26396.roadrunner.tuning;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.team26396.teamcode.MecanumDrive;
-import org.firstinspires.ftc.team26396.teamcode.TankDrive;
+import org.firstinspires.ftc.team26396.roadrunner.MecanumDrive;
+import org.firstinspires.ftc.team26396.roadrunner.TankDrive;
 
 public final class SplineTest extends LinearOpMode {
     @Override
